blogApp/requirements.txt
astroid==2.3.3
autopep8==1.4.4
boto3==1.15.7
botocore==1.18.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-boto==0.3.12
django-countries==6.1.2
django-crispy-forms==1.9.0
django-environ==0.4.5
django-js-asset==1.2.2
django-lint==2.0.4
django-mptt==0.11.0
django-storages==1.10.1
django-summernote==0.8.11.6
django-taggit==1.3.0
idna==2.9
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
logilab-common==1.5.2
mccabe==0.6.1
mysql-connector-python==8.0.20
oauthlib==3.1.0
Pillow==7.2.0
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
wrapt==1.11.2
